Former-commit-id: 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly f65e8a7253c65b5d59e78ed57e3d73bf01b2e06f]]]
Former-commit-id: 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee]
Former-commit-id: 29677326c7a0b9dcf115b06d8cb2e0645a2c7614
Former-commit-id: b6af57b411b9ec4c9fa3ba83714e67adf8c15db1

+++ edexOsgi/com.raytheon.edex.plugin.goessounding/src/com/raytheon/edex/plugin/goessounding/decoder/GOESSoundingDataAdapter.java
@@ -62,6 +62,7 @@ import com.raytheon.uf.edex.decodertools.core.IDecoderConstants;
  * May 09, 2013 1869       bsteffen    Modified D2D time series of point data to
  *                                     work without dataURI.
  * May 14, 2014 2536       bclement    moved WMO Header to common, removed TimeTools usage
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -188,7 +189,8 @@ public class GOESSoundingDataAdapter {
                     SPIEntry s = SPI_DATA.nearest(lat, lon, MAX_DISTANCE);
                     if (s != null) {
                         SurfaceObsLocation location = new SurfaceObsLocation();
-                        location.assignLocation(lat, lon);
+                        location.assignLocation(lat.floatValue(),
+                                lon.floatValue());
                         location.setStationId(s.getId());
                         obsData.setLocation(location);
                     } else {
